requirements.txt
pyautogen>=0.2.0
gitpython>=3.1.0
python-dotenv>=1.0.0
